modify ConstantNumberGenerator, generalize to ConstantGenerator (enforcing numbers not necessary) add ZipfianNumberGenerator, generic version of the YCSB ZipfianGenerator, which only returned Long add beginnings of GrowingRangeExponentialNumberGenerator (only test class so far)

+++ core/src/main/java/com/yahoo/ycsb/util/ByteArrayByteIterator.java
@@ -16,7 +16,6 @@
  */
 package com.yahoo.ycsb.util;
 
-
 public class ByteArrayByteIterator extends ByteIterator
 {
     byte[] str;
@@ -52,7 +51,7 @@ public class ByteArrayByteIterator extends ByteIterator
     }
 
     @Override
-    public long bytesLeft()
+    public int bytesLeft()
     {
         return len - off;
     }
